Use "cp" instead of "copy" so permissions are preserved
This fixes bug #910125.  Before, files copied by DH_GOLANG_INSTALL_EXTRA
did not maintain permissions.  This caused issues with some files that
had their execute bit stripped which was causing tests to fail.  The
difference between "cp" and "copy" in perl is that "cp" maintains
permissions.  According to the docs:

    You may use the syntax use File::Copy "cp" to get at the cp alias
    for this function. The syntax is exactly the same. The behavior is
    nearly the same as well: as of version 2.15, cp will preserve the source
    file's permission bits like the shell utility cp(1) would do, while copy
    uses the default permissions for the target file (which may depend on
    the process' umask, file ownership, inherited ACLs, etc.).